Add new tests
Test 'Image Content' should confirm that the server is able to serve
image content. Provide three images for the test with the extensions
'.gif' '.jpeg' and '.png'

Test 'Redirect Path' should confirm that the server sends a response
indicating a redirect.

Test 'Time To Complete' should verify that all tests are completing
successfully.